Move security related classes to a new package
