+++ src/main/java/com/sfjs/security/AuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.sfjs;
+package com.sfjs.security;
 
 import java.io.IOException;
 import java.io.PrintWriter;
