+++ src/main/java/com/sfjs/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.sfjs;
+package com.sfjs.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
